requirements.txt
dice==3.1.2
discord.py==1.7.3
Jinja2==3.1.2
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.1
tqdm==4.64.0
